pyproject.toml
[tool.poetry]
name = "ltd-invoice"
version = "0.1.0"
description = "Bookkepping automation"
authors = ["jonasborges <jonasborgesalves@gmail.com>"]
license = "unlicense"

[tool.poetry.dependencies]
python = "^3.9"
google-api-python-client = "^2.40.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.0"
"pdfminer.six" = "^20211012"
selenium = "^4.1.3"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"
pre-commit = "^2.17.0"
flake8 = "^4.0.1"
pytest-cov = "^3.0.0"
mypy = "^0.940"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
known_third_party = ["pdfminer", "google", "google_auth_oauthlib", "googleapiclient"]
known_first_party = ["ltd_invoice"]

[tool.mypy]
# --strict
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
ignore_missing_imports = true
implicit_reexport = false
no_implicit_optional = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true
# --strict end
exclude = ["^tests_\\.py$"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
